Tried to fix the bug that stopped dividing that led to irrational numbers.
Not completely happy with the results but It's a lot better than it was.
It's probably slow and it does strange things with the precision. But I
guess that's what you get when you use a fixed precision arithmetic. I
could say all calculation is done at 2 decimal places which would be
good with money but bad with other things or all 20 decimal places which
would be good for everything but display. Maybe I could do that but
change the display method to only show the first X digits where X is
defined in the settings. I could even get the precision to be in the
settings:

    display-precision = 2
    eval-precision = 20

I'll leave it as an optional.