fixes #3 The stand-alone jar has been tested on one PC.
The tests given in the comments for the issue have been used. This threw
up a bug (load overwrites parts of the history) which has been given a
fix(but it has a TODO for a better fix). It also fixes another bug with
reset not resetting the ID. 

I have added `commandLineJar.jardesc` which can be easily used to export
the program for distribution.